% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Sample Data Set}
\format{
## `example1`
A data frame with 85 rows and 697 columns:
\describe{
  \item{sampleType}{sample type}
  \item{time}{acquisition order}
  \item{label}{unique sample name}
  \item{Group}{treatment group}
  \item{metabolite_names}{compound names}
  ...
}
}
\source{
Stevens et. al. 2023
}
\usage{
example1
}
\description{
A data matrix from Stevens et. al. 2023 containing LC-MS/MS based
untargeted metabolomics and lipidomics data.
}
\keyword{datasets}
