% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.params.R
\name{ms_add_anno}
\alias{ms_add_anno}
\title{Add annotations to lipid database}
\usage{
ms_add_anno(f)
}
\arguments{
\item{f}{Name of a data file in the "data/" folder.
Accepts .xlsx, .csv, and .txt files.}

\item{md_num}{Number of metadata columns in input data file.}

\item{qc_rep}{Should a QC report be generated for the selected dataset?
(Either TRUE or FALSE)}

\item{norm1}{Logical indicating if the data have been normalized.
(Either TRUE or FALSE)}

\item{ref1}{Reference annotation list for assigning compound details and
class information (provided as a .txt file).}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Appends annotations from a dataset not already present in the
lipid database.
}
\examples{

# ms_input(
#   "example.txt",
#   4,
#   FALSE,
#   TRUE,
#   "reference.txt"
# )

}
