% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.check.lib.R
\name{ms_check_lib}
\alias{ms_check_lib}
\title{Install Libraries}
\usage{
ms_check_lib()
}
\value{
Summary of installed packages
}
\description{
Installs CRAN and Bioconductor dependencies
}
\examples{

 # ms_check_lib()

}
