% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.data.sum.ann.R
\name{ms_plot_anno}
\alias{ms_plot_anno}
\title{Annotation Summary}
\usage{
ms_plot_anno(ld, md)
}
\arguments{
\item{ld}{A data list containing annotations assigned by ms_input().}

\item{md}{A vector of annotation levels to display in the plot.}
}
\value{
List containing a formatted annotation table and sunburst plot
summarizing annotation information.
}
\description{
Summarizes annotation information for a specific dataset.
}
\examples{

# ms_plot_anno(
#   d,
#   c("Major.Class", "Class", "Subclass")
# )

}
