% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.chemrich.R
\name{ms_plot_crich}
\alias{ms_plot_crich}
\title{Plot ChemRICH Results}
\usage{
ms_plot_crich(dfcr)
}
\arguments{
\item{dfcr}{An individual comparison from a ChemRICH results data frame.}
}
\value{
A dot plot displaying ChemRICH results,
including cluster size, increased/decreased proportion,
and statistical significance.
}
\description{
Plots ChemRICH results for selected comparison as a dot plot.
}
\examples{
# ms_plot_crich(
#   dfcr = d[["chemrich"]][d[["chemrich"]][["Comparison"]] == "comp1", ],
#   p_title = "comp1"
# )

}
