% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.visualize.R
\name{ms_plot_heat}
\alias{ms_plot_heat}
\title{Heatmap 1}
\usage{
ms_plot_heat(
  d_stat,
  d_ref,
  c_list,
  diff_col,
  p_col,
  an1,
  hm_w,
  hm_h,
  fs_r,
  cl_c,
  cl_r
)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_anova().}

\item{d_ref}{Reference annotation list generated by ms_input().}

\item{c_list}{Comparison list, provided as a vector.}

\item{diff_col}{Column name containing effect size.}

\item{p_col}{Column name containing adjusted p-values.}

\item{an1}{Variable names used for annotating the heatmap.}

\item{hm_w}{Heatmap width.}

\item{hm_h}{Heatmap height.}

\item{fs_r}{Row fontsize.}

\item{cl_c}{Logical indicating if the columns should be clustered.}

\item{cl_r}{Logical indicating if the rows should be clustered.}
}
\value{
A heatmap for the selected comparisons.
}
\description{
Generates a heatmap from a statistical results object for
a vector of comparisons.
}
\examples{

# ms_plot_heat(
#   d_stat = d[["stats"]],
#   d_ref = d[["data"]][["anno"]],
#   c_list = c("KO vs. Control"),
#   diff_col = "Log2FC",
#   p_col = "FDR",
#   an1 = c("Major.Class", "Saturation"),
#   hm_w = 36,
#   hm_h = 10,
#   fs_r = 8,
#   cl_c = TRUE,
#   cl_r = TRUE
# )

}
