% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.visualize.R
\name{ms_plot_lipidnet}
\alias{ms_plot_lipidnet}
\title{Lipid Network Plot}
\usage{
ms_plot_lipidnet(d_enrch, net_type)
}
\arguments{
\item{d_enrch}{Enrichment results list generated by ms_stat_crich().
Only required when network type is "annotation" or "enrichment".}

\item{net_type}{Output network type. Currently supported types are
"base", "annotation", or "enrichment".}
}
\value{
A network plot of the specified type.
}
\description{
Generates a network plot using known lipid metabolism pathways
to visualize class-based enrichment statistics. Uses a manually
curated collection of pathways as a reference. Lipid class
annotations assigned by this reference enable mapping of any
untargeted or targeted lipidomics dataset onto the network.
}
\examples{

# ms_plot_lipidnet(
#   d_enrch = d[["chemrich"]],
#   net_type = "enrichment",
#
#   c_list = c("KO vs. Control"),
#   diff_col = "Log2FC",
#   p_col = "FDR",
#   an1 = c("Major.Class", "Saturation"),
#   hm_w = 36,
#   hm_h = 10,
#   fs_r = 8,
#   cl_c = TRUE,
#   cl_r = TRUE
# )

}
