% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.visualize.R
\name{ms_plot_volcano}
\alias{ms_plot_volcano}
\title{Volcano Plot}
\usage{
ms_plot_volcano(
  d_stat,
  comp_name,
  diff_col,
  p_col,
  p_cut,
  f_cut,
  f_lim,
  y_limit,
  x_title
)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_anova().}

\item{comp_name}{Comparison name, provided as a character string.}

\item{diff_col}{Column name containing effect size.}

\item{p_col}{Column name containing adjusted p-values.}

\item{p_cut}{Numeric value for p-value cutoff to indicate significance.}

\item{f_cut}{Fold change cutoff indicating high effect size.}

\item{f_lim}{Numeric value for fold change limits on the x-axis.}

\item{y_limit}{Upper-bound -log10 p-value to display on the y-axis.}

\item{x_title}{X-axis legend title, provided as a character string.}
}
\value{
A volcano plot for the chosen treatment comparisons.
}
\description{
Generates a volcano plot from a statistical results object for
a chosen comparison.
}
\examples{

# ms_plot_volcano(
#   d_stat = d[["stats"]],
#   comp_name = "KO vs. Control",
#   diff_col = "Log2FC",
#   p_col = "p adj",
#   p_cut = 0.05,
#   f_cut = 0.25,
#   f_lim = 6,
#   y_limit = 10,
#   x_title = "x-axis title"
# )

}
