% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.stats.univar.R
\name{ms_stat_anova}
\alias{ms_stat_anova}
\title{Multivariate ANOVA}
\usage{
ms_stat_anova(
  matpv,
  matfc,
  md,
  md_var,
  an,
  fc_class1 = FALSE,
  grp_class1 = NULL
)
}
\arguments{
\item{matpv}{Data matrix for conducting ANOVA.}

\item{matfc}{Data matrix for calculating fold change.}

\item{md}{Metadata from ms_input().}

\item{md_var}{Metadata variables to test by ANOVA.}

\item{an}{Reference annotation data frame.}

\item{fc_class1}{Logical indicating if compound classes rather
than individual compounds should be tested for significant differences.}

\item{grp_class1}{If fc_class1 is TRUE, which variable should be used
to group individual compounds into separate compound classes?}
}
\value{
A data frame containing the combined ANOVA and fold change results
for the specified variable(s).
}
\description{
Conducts an ANOVA for each metadata variable with Tukey's post-hoc
analysis for each compound and combines with fold changes
for each group comparison.
}
\examples{

# ms_stat_anova(
#   matpv = d[["data"]][["data.scale"]],
#   matfc = d[["data"]][["imputed"]],
#   md = d[["data"]][["meta"]],
#   md_var = c("Organ", "Sex", "Dose"),
#   an = d[["data"]][["anno"]],
#   fc_class1 = FALSE,
#   grp_class1 = "Subclass"
# )

}
