% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.chemrich.R
\name{ms_stat_crich}
\alias{ms_stat_crich}
\title{ChemRICH Analysis Function}
\usage{
ms_stat_crich(d_stat, d_ref, cl_name)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_anova().}

\item{d_ref}{Data frame to use for assigning annotation classes.}

\item{cl_name}{Variable name to use for assigning classes and grouping
for enrichment analysis.}
}
\value{
A ChemRICH results data frame.
}
\description{
Conducts ChemRICH analysis for each comparison in a provided result
generated by ms_stat_anova() (default) or from a single
ChemRICH input (in development). The original enrichment method is adapted
from Chemical Similarity Enrichment Analysis (ChemRICH), which is detailed
in Barupal et. al. 2017 (DOI: 10.1038/s41598-017-15231-w).
Significantly enriched classes are determined by a weighted
Kolmogorov-Smirnov Test adjusted for false discovery rate (FDR).
}
\examples{
# ms_stat_crich(
#   d_stat = d[["stats"]],
#   d_ref = d[["data"]][["anno"]],
#   cl_name = "label.saturation"
# )

}
