% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.fc.R
\name{ms_stat_fc}
\alias{ms_stat_fc}
\title{Group-wise Fold Change Function}
\usage{
ms_stat_fc(mat1, md, md_var, an, fc_class, grp_class)
}
\arguments{
\item{mat1}{Data matrix assigned by ms_input().}

\item{md}{Dataframe containing study metadata.}

\item{md_var}{Vector of metadata variables to calculate fold
changes for.}

\item{an}{Dataframe containing annotation metadata.}

\item{fc_class}{Logical indicating whether class-wide
fold changes should be calculated. Requires a valid
annotation list contained within ms_input() result.}

\item{grp_class}{Name of the class annotation column
to use. Ignored when fc_class is set to FALSE.}
}
\value{
A data frame containing the group-wise
fold changes for each compound.
}
\description{
Calculates the log2 fold changes for all
possible combinations of selected metadata variables.
}
\examples{

# ms_stat_fc(
#   mat1 = d[["data"]][["imputed"]],
#   md = d[["data"]][["meta"]],
#   md_var = c("Organ", "Sex", "Dose"),
#   an = d[["data"]][["anno"]],
#   fc_class = TRUE,
#   grp_class = "Subclass"
# )

}
