% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{ms_theme_net}
\alias{ms_theme_net}
\title{Network Plot Theme}
\usage{
ms_theme_net()
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
General network plot theme.
}
\examples{

# ms_theme_net()

}
