% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{ms_theme_net2}
\alias{ms_theme_net2}
\title{Network Plot Theme (Alternate)}
\usage{
ms_theme_net2()
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
Alternate network plot theme.
}
\examples{

# ms_theme_net2()

}
