% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.data.trf.sum.R
\name{ms_data_check}
\alias{ms_data_check}
\title{Data Transformation and Quality Check}
\usage{
ms_data_check(ld)
}
\arguments{
\item{ld}{Data list generated from ms_input().}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Performs data imputation, log2-transformation, and
quality checks on an untargeted dataset.
}
\examples{

# ms_data_check(d)

}
