% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.visualize.R
\name{ms_plot_lipidnet}
\alias{ms_plot_lipidnet}
\title{Lipid Network Plot}
\usage{
ms_plot_lipidnet(
  d_enrch = NULL,
  net_type,
  cl_var = NULL,
  comp1 = NULL,
  comp_sat = NULL
)
}
\arguments{
\item{d_enrch}{Enrichment results list generated by ms_stat_crich().
Only required when network type is "annotation" or "enrichment".}

\item{net_type}{Output network type. Currently supported types are
"pathways_legend", "base", "annotation", or "enrichment".}

\item{cl_var}{Clustering variable to use for aggregating lipid classes.}

\item{comp1}{Comparison name for plotting enrichment statistics. Currently
supports up to one comparison name.}

\item{comp_sat}{Should unsaturated and saturated lipid classes be compared?
Enter a logical value (TRUE/FALSE) to toggle lipid saturation comparisons.
This parameter is dependent upon the presence of both unsaturated and
saturated lipid classes in the enrichment statistics data frame.}
}
\value{
A network plot of the specified type.
}
\description{
Generates a network plot using known lipid metabolism pathways
to visualize class-based enrichment statistics. Uses a manually
curated collection of pathways as a reference. Lipid class
annotations assigned by this reference enable mapping of any
untargeted or targeted lipidomics dataset onto the network.
}
\examples{

# ## Pathway Legend
# ms_plot_lipidnet(
#   net_type = "pathways_legend"
# )
# ## Base network
# ms_plot_lipidnet(
#   net_type = "base"
# )
# ## Plot dataset annotations
# ms_plot_lipidnet(
#   d_enrch = d[["chemrich"]],
#   net_type = "annotation",
#   cl_var = "label_saturation"
# )
# ## Plot dataset enrichment results
# ms_plot_lipidnet(
#   d_enrch = d[["chemrich"]],
#   net_type = "enrichment",
#   cl_var = "label_saturation",
#   comp1 = "HFA_F-BR",
#   comp_sat = TRUE
# )

}
